requirements.txt
streamlit
openai
python-dotenv
scapy
netifaces